Update tests location to match the source file name.
