+++ tests/number.test.ts
@@ -1,6 +1,150 @@
-import { range } from "../source/utilities";
+import {
+    range,
+    getRandomFloat,
+    getRandomInt,
+    getSeveralRandomInts,
+    numberOfDigits,
+    round,
+} from "../source/utilities";
 
-describe("range()", () => {
+describe("getRandomFloat", () => {
+    test("Test with valid arguments.", () => {
+        var ok;
+
+        // inclusive limits
+        var result = getRandomFloat(2, 2);
+        expect(result).toBe(2);
+
+        // a random value between the limits
+        result = getRandomFloat(-4.44, 4.44);
+        ok = result >= -4.44 && result <= 4.44;
+        expect(ok).toBe(true);
+
+        // zero in one of the range limits
+        result = getRandomFloat(0, 4.123);
+        ok = result >= 0 && result <= 4.123;
+
+        expect(ok).toBe(true);
+    });
+});
+
+describe("getRandomInt", () => {
+    test("Test with valid arguments.", () => {
+        // inclusive limits
+        var result = getRandomInt(4, 4);
+        expect(result).toBe(4);
+
+        // a value between -4 and 4
+        result = getRandomInt(-4, 4);
+        var within = result >= -4 && result <= 4;
+
+        expect(within).toBe(true);
+
+        // zero in one of the range limits
+        result = getRandomInt(0, 2);
+        within = result >= 0 && result <= 2;
+
+        expect(within).toBe(true);
+    });
+});
+
+describe("getSeveralRandomInts", () => {
+    test("Test with valid arguments.", () => {
+        var ok;
+        var result = getSeveralRandomInts(1, 1, 1);
+
+        // range of 1 with 1 random integer needed
+        expect(result).toStrictEqual([1]);
+
+        // a range of 2, with 2 required integers
+        result = getSeveralRandomInts(1, 2, 2);
+        var range = [1, 2];
+
+        ok = true;
+
+        if (result.length === range.length) {
+            for (let a = 0; a < range.length; a++) {
+                if (result.indexOf(range[a]) < 0) {
+                    ok = false;
+                    break;
+                }
+            }
+        } else {
+            ok = false;
+        }
+
+        expect(ok).toBe(true);
+
+        // a range bigger than the number of required integers
+        result = getSeveralRandomInts(-4, 4, 4);
+        ok = true;
+
+        if (result.length === 4) {
+            for (let a = 0; a < result.length; a++) {
+                var value = result[a];
+
+                if (value < -4 || value > 4) {
+                    ok = false;
+                    break;
+                }
+            }
+        } else {
+            ok = false;
+        }
+
+        expect(ok).toBe(true);
+
+        // a zero in the range limits
+        result = getSeveralRandomInts(0, 0, 1);
+        ok = result[0] === 0;
+
+        expect(ok).toBe(true);
+    });
+
+    test("Should include at least once each possible value, if the 'howMany' is higher than the range.", () => {
+        const result = getSeveralRandomInts(-5, 5, 20);
+
+        expect(result.length).toBe(20);
+
+        range(-5, 5).forEach((value) => {
+            expect(result.includes(value)).toBeTruthy();
+        });
+    });
+});
+
+describe("numberOfDigits", () => {
+    test("Test with valid arguments.", () => {
+        var result = numberOfDigits(4);
+        expect(result).toBe(1);
+
+        result = numberOfDigits(12);
+        expect(result).toBe(2);
+
+        result = numberOfDigits(-12);
+        expect(result).toBe(2);
+
+        result = numberOfDigits(12.11);
+        expect(result).toBe(4);
+
+        result = numberOfDigits(-12.34);
+        expect(result).toBe(4);
+    });
+});
+
+describe("round", () => {
+    test("Test with valid arguments.", () => {
+        var result = round(4.22, 1);
+        expect(result).toBe(4.2);
+
+        result = round(4.25, 1);
+        expect(result).toBe(4.3);
+
+        result = round(-4.5, 0);
+        expect(result).toBe(-4);
+    });
+});
+
+describe("range", () => {
     test("Should create a list with the numbers in-between the given arguments.", () => {
         const result = range(0, 10);
